renaming module ml_vectorizer -> static_vectorizer
