+++ metalibm_core/core/static_vectorizer.py
