+++ metalibm_core/targets/common/dummy_asm_target.py
@@ -70,7 +70,7 @@ from metalibm_core.core.bb_operations import SequentialBlock
 from metalibm_core.core.machine_operations import (
     MaterializeConstant, RegisterCopy, RegisterAssign, SubRegister, VirtualRegister)
 from metalibm_core.core.ml_complex_formats import is_pointer_format, ML_Pointer_Format
-from metalibm_core.core.ml_vectorizer import vectorize_format
+from metalibm_core.core.static_vectorizer import vectorize_format
 
 from metalibm_core.code_generation.generator_utility import (
     TemplateOperatorFormat,
